requirements.txt
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
langchain>=0.1.0
google-generativeai>=0.3.0
numpy>=1.20.0
requests>=2.25.0
streamlit>=1.20.0